[ADD] account_bank_balance_report: Migrated to version 8.0 with qweb report
