+++ account_bank_balance_report/report/report.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+     <data>
+        
+        <!-- QWeb Reports -->
+        <report id="action_report_bank_balance" 
+        model="account.account"
+        string="Bank Balance Report" 
+        report_type="qweb-pdf"
+        name="account_bank_balance_report.report_bank_balance" 
+        file="account_bank_balance_report.report_bank_balance"
+        menu="False"
+        />
+    </data>
+</openerp>
